client/requirements.txt
gradio==5.38.0
python-dotenv==1.1.1
websockets==15.0.1
